Second edition completed!

+++ controllers/BlogController.php
@@ -8,6 +8,7 @@ use app\core\View;
 use app\models\Blog;
 use app\models\SearchContext;
 use app\models\Search;
+use mysql_xdevapi\Result;
 
 class BlogController
 {
@@ -18,54 +19,38 @@ class BlogController
         $this->blog = new Blog();
     }
 
-    protected function checkId($id)
-    {
-        $blog = $this->blog->findOne(['id' => ['=', $id]]);
-
-        if (!$blog)
-            throw new NotFoundException();
-
-        return $blog;
-    }
+//    protected function checkId($id)
+//    {
+//        $blog = $this->blog->findOne(['id' => ['=', $id]]);
+//
+//        if (!$blog)
+//            throw new NotFoundException();
+//
+//        return $blog;
+//    }
 
     public function index($params)
     {
-        // Search
         $search = new Search();
         $search->loadData(App::request()->getBody());
 
-        // Pagination
-        $currentPage = $params['page'] ?? 1;
-        if (!is_numeric($currentPage) || $currentPage < 1)
-            $currentPage = 1;
-        $currentPage = intval($currentPage);
-
-        $limit = 5;
-        // Count all records
-        $allRecords = $this->blog->count([
-            'title' => ['LIKE' , $search->search, '%', '%'],
-        ]);
-
-        $blogs = $this->blog->paginate($currentPage, $limit, [
-            'title' => ['LIKE' , $search->search, '%', '%'],
+        $paginate = $this->blog->paginate($params['page'] ?? 1, 5, [
+            ['title', 'context', 'name'], $search->search,
         ]);
 
         return new View('blog/index', [
-            'blogs' => $blogs,
+            'paginate' => $paginate,
             'search' => $search,
-            'allRecords' => $allRecords,
-            'currentPage' => $currentPage,
-            'limit' => $limit,
         ]);
     }
 
     public function show($params)
     {
         $id = $params['id'];
-        $blog = $this->checkId($id);
+        $this->blog->checkId($id);
 
         return new View('blog/show', [
-            'blog' => $blog,
+            'blog' => $this->blog,
             'id' => $id,
         ]);
     }
@@ -85,24 +70,16 @@ class BlogController
                 'blog' => $this->blog,
             ]);
 
-        $this->blog->save();
-
-        App::session()->setFlash('success', 'Blog created successfully!');
-        App::response()->redirect('/blog');
-        echo 111;
-        exit();
-        return true;
+        return $this->blog->save();
     }
 
     public function edit($params)
     {
-        $id = $params['id'];
-        $blog = $this->checkId($id);
+        if (!$this->blog->checkUserAccess())
+            return false;
 
-        $this->blog->id = $blog->id;
-        $this->blog->title = $blog->title;
-        $this->blog->author = $blog->author;
-        $this->blog->context = $blog->context;
+        $id = $params['id'];
+        $this->blog->checkId($id);
 
         return new View('blog/edit', [
             'blog' => $this->blog,
@@ -112,8 +89,11 @@ class BlogController
 
     public function update($params)
     {
+        if (!$this->blog->checkUserAccess())
+            return false;
+
         $id = $params['id'];
-        $this->checkId($id);
+        $this->blog->checkId($id);
 
         $this->blog->loadData(App::request()->getBody());
         if (!$this->blog->validation())
@@ -122,22 +102,17 @@ class BlogController
                 'id' => $id,
             ]);
 
-        $this->blog->update($id);
-
-        App::session()->setFlash('success', 'Blog edited successfully!');
-        App::response()->redirect('/blog');
-        return true;
+        return $this->blog->update($id);
     }
 
     public function destroy($params)
     {
+        if (!$this->blog->checkUserAccess())
+            return false;
+
         $id = $params['id'];
-        $blog = $this->checkId($id);
+        $this->blog->checkId($id);
 
-        $this->blog->destroy($id);
-
-        App::session()->setFlash('success', 'Blog deleted successfully!');
-        App::response()->redirect('/blog');
-        return true;
+        return $this->blog->destroy($id);
     }
 }
